requirements.txt
python-telegram-bot[ext]==20.7
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
APScheduler==3.10.4
python-dotenv==1.0.1
